Multiplayer Intractability Patch
The first multiplayer intractable item has been introduced to the game in the form of a red cube in the test area. This cube can be interacted with by all players by transferring ownership of the cube once picked up. This will not be done for all items in the game to reduce latency but the necessary items will have a multiplayer intractability aspect