Skeld Map Organisation Patch #1
This patch is the first of 3 patches that will reorganize the Skeld map in the Room 2 lobby, increasing the ease of use when implementing tasks into the map. This first patch begins by recreating the map and applying naming standards that easier to follow, this is  dont to all of the rooms and hallways, with the patch, the vents of the map are missing, with will arrive in the next version of this patch.